+++ bin/generis
@@ -4,14 +4,17 @@
 
 require_once __DIR__ . '/../vendor/autoload.php';
 
-$commands = include __DIR__ . '/../config/generis/commands.conf.php';
-
+use oat\generis\Model\Config\Reader;
 use Symfony\Component\Console\Application;
+use oat\generis\Model\DependencyInjection\ObjectFactory;
 
 $app = new Application();
+$configReader = new Reader();
+$objectFactory = new ObjectFactory();
 
+$commands = (array) $configReader->get('generis/commands');
 foreach ($commands as $command) {
-    $app->add($command);
+    $app->add($objectFactory->create($command));
 }
 
 $app->run();
