--- config/extension.yml
@@ -1,2 +1,0 @@
-generis:
-  namespace: '\oat\generis'
